- Method getAll() was added to UserService.  - Method update in UserService accepts two arguments now - id and user.  - UserController was added. There are CRUD methods in it.  - Tests classes were distributed on packages.  - To the test.java package was added ObjectMapping class for transforming objects to and from json.  - Class CRUDUserControllerTests was added. It contains tests for UserController.

+++ src/main/java/phonebooks/services/UserService.java
@@ -5,6 +5,8 @@ import org.springframework.transaction.annotation.Transactional;
 import phonebooks.entities.User;
 import phonebooks.repositories.UserRepository;
 
+import java.util.List;
+
 @Service
 @Transactional
 public class UserService {
@@ -18,16 +20,31 @@ public class UserService {
         return repository.save(user);
     }
 
+    public List<User> getAll(){
+        return repository.findAll();
+    }
+
     public User getById(Long id) {
         return repository.findById(id).orElse(null);
     }
 
-    public User update(User user) {
-        return repository.save(user);
+    public User update(Long id, User user) {
+        User savedUser = repository.findById(id).orElse(null);
+        if(savedUser != null) {
+            user.setId(savedUser.getId());
+            return repository.save(user);
+        } else {
+            return null;
+        }
     }
 
-    public void deleteById(Long id) {
-        repository.deleteById(id);
+    public boolean deleteById(Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
 }
